refactor(core): Frame strong type — struct replacing bare i64 alias (Phase 1)
frame.hpp: struct Frame{i64 value} with constexpr ctor, spaceship, arithmetic, explicit conversions, named accessors, ostream, fmt::formatter outside namespace

time.hpp: remove duplicate using Frame=i64, add frame.hpp include

hash_builder.hpp: add add(Frame) overload (struct no longer integral)

render_job_write_frame.cpp: frame.as_i64() for format_path

command_camera_path.cpp: static_cast<int>(Frame) for 3 int params

designated init fixes: reorder .color/.opacity_easing in text_animations.cpp + cinematic_text_camera.cpp

src/CMakeLists.txt: re-remove Taskflow::Taskflow (undone by stash)

+++ apps/chronon3d_cli/commands/dev/command_camera_path.cpp
@@ -197,7 +197,7 @@ int command_camera_path(const CompositionRegistry& registry,
 
         const auto& cam = scene.camera_2_5d();
         PathSample s;
-        s.frame      = f;
+        s.frame      = static_cast<int>(f);
         s.position   = cam.position;
         s.rotation   = cam.rotation;
         s.zoom       = cam.zoom;
@@ -219,7 +219,7 @@ int command_camera_path(const CompositionRegistry& registry,
     compute_metrics(samples, metrics);
 
     // ── Print terminal summary ──────────────────────────────────────────────
-    print_summary(args.comp_id, metrics, start, end, step);
+    print_summary(args.comp_id, metrics, static_cast<int>(start), static_cast<int>(end), step);
 
     // ── Export if output path specified ──────────────────────────────────────
     if (!args.output.empty()) {
@@ -233,7 +233,7 @@ int command_camera_path(const CompositionRegistry& registry,
         if (fmt_str == "csv") {
             ok = write_csv(args.output, samples);
         } else {
-            ok = write_json(args.output, args.comp_id, start, end, step, samples, metrics);
+            ok = write_json(args.output, args.comp_id, static_cast<int>(start), static_cast<int>(end), step, samples, metrics);
         }
 
         if (ok) {
